Kontena::Cli should be in lib/kontena/cli.rb, Kontena should be in lib/kontena.rb
